requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
transformers==4.36.0
torch==2.1.0
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1
numpy==1.24.3
scikit-learn==1.3.2